Added some more datasets and ranamed some files
